requirements.txt
pandas>=2.1.0
openpyxl>=3.1.0
flask>=3.0.0
Pillow>=10.1.0
watchdog>=3.0.0
python-dotenv>=1.0.0
requests>=2.28.0
aiogram>=3.12.0
aiohttp>=3.8.0
waitress>=3.0.0
